api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
openai==1.40.6
numpy==1.26.4
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
PyYAML==6.0.2
